# api/openapi.yaml
openapi: 3.1.0
info:
  title: Curio Daemon API
  version: 0.1.0
  description: |
    HTTP+JSON API exposed by `curio-daemon` on `127.0.0.1:8765` (the port is
    configurable; the host must be loopback). Consumed by the `curio` CLI,
    the `curio-mcp` sidecar, and any future client.

    This file is the contract. Tests in `internal/api` check it against the
    router (every route is documented, every documented operation is
    routed) and against live responses (every operation's real answers
    validate against their schemas, with undeclared fields treated as
    errors). The clients in `internal/client` are written by hand; code
    generation is deferred.

    Conventions:
      - IDs are UUIDs. Timestamps are RFC 3339, UTC.
      - Unset optional fields are omitted; the API never sends `null`.
      - Errors follow RFC 7807 (`application/problem+json`). Every response
        carries an `X-Request-Id` header, and every problem repeats it as
        `request_id`, which is also on the daemon's log lines for the
        request; `instance` is the request path. A 404 names what is
        missing. A 500's detail is the raw error text.
      - Responses are read tolerantly: they may gain fields and enum values
        within /v1, and clients ignore what they don't know. Requests are
        strict: an unknown field is a 400 that names it and, since it
        usually means a newer client talking to an older daemon, says to
...
        again.
...

...

...
    this API as JSON.
servers:
...
paths:
...
  /v1/documents/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
    get:
      tags: [documents]
      operationId: getDocument
      summary: Get document metadata and its current extraction
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Document" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }
...
  /v1/search:
    post:
      tags: [search]
      operationId: search
      summary: Hybrid BM25 + vector search
      description: >
        Runs BM25 and vector retrieval concurrently and fuses them with RRF.
        Failed and dead documents are not searched: they may keep the chunks
        of an earlier fetch, which no longer describe what the URL serves. A
        pending document (a refetch in flight) is. If the embedding model is
        unavailable (Ollama down, or slower than search.embed_timeout_seconds),
        the keyword results are still returned with degraded=true and a
        warning explaining why; a keyword-index failure is a 500.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SearchRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SearchResponse" }
        "400":
          description: Missing query, k outside 0..100, an unknown field, or a malformed body
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        default: { $ref: "#/components/responses/Problem" }

  # ---------- insight ----------
  /v1/interests:
    get:
      tags: [insight]
      operationId: listInterests
      summary: List inferred interests (labeled topic clusters)
      description: >
        Returns the interests from the latest completed clustering run, largest
        first, with run metadata. An interest is a labeled cluster. Returns an
        empty list when clustering has not run yet.
      parameters:
        - in: query
          name: limit
          description: Max interests. Out of range or malformed means the default.
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - in: query
          name: members
          description: Member documents per interest; 0 for none. Out of range or malformed means the default.
          schema: { type: integer, minimum: 0, maximum: 100, default: 5 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InterestList" }
        default: { $ref: "#/components/responses/Problem" }
...
  /v1/interests/{id}:
    parameters:
      - { $ref: "#/components/parameters/ID" }
      - in: query
        name: members
        description: Member documents to include; 0 for none. Out of range or malformed means the default.
        schema: { type: integer, minimum: 0, maximum: 1000, default: 100 }
    get:
      tags: [insight]
      operationId: getInterest
      summary: Get one interest with its member documents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Interest" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Problem" }

  # ---------- jobs ----------
...
components:
  parameters:
    ID:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: >
        An error. Besides the statuses an operation lists: 403 for a request
        a browser could have made (see the access policy), 405 for a method
        the path doesn't take (with an Allow header), 413 for an oversized
        body, 415 for a body that isn't application/json, 500 for a failure
        in the daemon, logged under the request_id, and 503 with a
        Retry-After header while the daemon is starting (type
        urn:curio:problem:daemon-starting; see the Startup convention).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: No such resource; the detail names it.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      description: RFC 7807 problem details
      type: object
      properties:
        type:       { type: string, format: uri, default: about:blank }
        title:      { type: string }
        status:     { type: integer }
        detail:     { type: string }
        instance:   { type: string, description: "The request path" }
        request_id: { type: string, description: "The response's X-Request-Id; quote it against the daemon log" }
      required: [title, status]
...
    DocState:
      type: string
      enum: [pending, fetched, failed, dead]
    ContentType:
      type: string
      enum: [article, repo, video, pdf, thread, unknown]
...
    Source:
      type: string
      enum: [chrome, safari, firefox, html, manual]

    # ---------- system ----------
...
    Document:
      type: object
      properties:
        id:                 { type: string, format: uuid }
        url:                { type: string, format: uri }
        url_canonical:      { type: string, format: uri }
        content_type:       { $ref: "#/components/schemas/ContentType" }
        title:              { type: string }
        author:             { type: string }
        published_at:       { type: string, format: date-time }
        language:           { type: string }
        word_count:         { type: integer }
        state:              { $ref: "#/components/schemas/DocState" }
        current_extraction: { $ref: "#/components/schemas/Extraction" }
        created_at:         { type: string, format: date-time }
        updated_at:         { type: string, format: date-time }
      required: [id, url, content_type, state, created_at, updated_at]
    Extraction:
      type: object
      properties:
        id:              { type: string, format: uuid }
        fetched_at:      { type: string, format: date-time }
        fetcher:         { type: string }
        status:          { type: string, enum: [ok, partial, paywalled, error] }
        markdown_path:   { type: string, description: "Absolute on-disk path to the extracted markdown" }
        error_message:   { type: string, description: "Why the extraction isn't ok. For a partial one, why its primary content is missing (a video transcript yt-dlp couldn't download, say)" }
        extraction_meta: { type: object, additionalProperties: true, description: "What the fetcher recorded; its shape varies by fetcher" }
      required: [id, fetched_at, fetcher, status]
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Natural-language or keyword query
        k:
          type: integer
          minimum: 0
          maximum: 100
          description: >
            Max documents to return after RRF fusion + collapse, 1 to 100.
            Omitted (or 0) uses the server's search.default_k (10 unless
            configured).
        filters:
          type: object
          description: >
            Every dimension given is applied. Folder and tag filters,
            per-request RRF weights, collapse strategy and saved-date filters
            are not implemented; the server config sets the weights and
            collapse strategy, and sending any of them is a 400.
          properties:
            content_type: { type: array, items: { $ref: "#/components/schemas/ContentType" } }
            host:         { type: array, items: { type: string }, description: "URL hosts, matched exactly and case-insensitively" }
            source:       { type: array, items: { $ref: "#/components/schemas/Source" } }
      required: [query]
    SearchHit:
      type: object
      properties:
        document:      { $ref: "#/components/schemas/Document" }
        score:         { type: number, description: "Fused RRF score for search, vector similarity for related (higher = better)" }
        markdown_path: { type: string, description: "Absolute on-disk path to the document's current markdown" }
        matches:
          type: array
          description: Top matching chunks for this document, in score order
          items: { $ref: "#/components/schemas/ChunkMatch" }
      required: [document, score]
    ChunkMatch:
      type: object
      properties:
        chunk_id:     { type: string, format: uuid }
        text:         { type: string, description: "The chunk's text" }
        snippet:      { type: string, description: "The BM25 snippet, with <em> highlight markers; absent for vector-only matches" }
        bm25_score:   { type: number, description: "Absent when BM25 didn't return the chunk" }
        vector_score: { type: number, description: "Absent when vector search didn't return the chunk" }
      required: [chunk_id, text]
    SearchResponse:
      type: object
      properties:
        query:       { type: string }
        took_ms:     { type: integer, description: "Time the search took, in milliseconds" }
        bm25_hits:   { type: integer, description: "How many chunks BM25 returned before fusion" }
        vector_hits: { type: integer, description: "How many chunks the vector index returned before fusion (0 when degraded)" }
        degraded:
          type: boolean
          description: >
            Present and true when semantic search was unavailable and items
            are keyword-only (BM25) results.
        warnings:
          type: array
          items: { type: string }
          description: Why the results are degraded; present only when there is something to report.
        items:
          type: array
          items: { $ref: "#/components/schemas/SearchHit" }
      required: [query, took_ms, bm25_hits, vector_hits, items]

    # ---------- insight ----------
    InterestMember:
      type: object
      properties:
        doc_id:        { type: string, format: uuid }
        title:         { type: string }
        url:           { type: string, format: uri }
        markdown_path: { type: string, description: "Absolute on-disk path to the member's extracted markdown" }
        similarity:    { type: number, description: "Cosine similarity to the cluster centroid (0..1)" }
      required: [doc_id, url, similarity]
    Interest:
      type: object
      description: A labeled topic cluster.
      properties:
        id:       { type: string, format: uuid }
        label:    { type: string, description: "Topic name; omitted when unlabeled" }
        summary:  { type: string }
        size:     { type: integer, description: "Number of member documents" }
        cohesion: { type: number, description: "Mean member cosine to the cluster centroid (0..1)" }
        members:
          type: array
          items: { $ref: "#/components/schemas/InterestMember" }
      required: [id, size, cohesion]
    InterestList:
      type: object
      description: >
        Not cursor-paginated: interests are a small bounded snapshot returned
        with the metadata of the run that produced them. The run fields are
        absent before the first clustering run.
      properties:
        run_id:        { type: string, format: uuid }
        computed_at:   { type: string, format: date-time }
        algo:          { type: string, description: "Clusterer name, e.g. knn-graph" }
        num_documents: { type: integer, description: "Documents considered in the run" }
        num_clusters:  { type: integer }
        num_noise:     { type: integer, description: "Documents left unclustered" }
        items:
          type: array
          items: { $ref: "#/components/schemas/Interest" }
      required: [num_documents, num_clusters, num_noise, items]

    # ---------- jobs ----------
...